# apps/web/openapi.codegen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sahabatkreator API (codegen)",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/v1/accounts/{platform}/authorize": {
      "get": {
        "summary": "Mulai OAuth untuk menghubungkan akun",
        "description": "Mengembalikan URL otorisasi platform. Setelah pengguna menyetujui, kami mengalihkan browser ke `redirect` Anda dengan `code` + `state` (atau `error` bila ditolak), lalu Anda menukarkan `code` lewat `POST /v1/accounts/{platform}/connect`.\n\nButuh plan Business+ (`api_write`) dan API key yang terikat ke developer app.",
        "tags": ["Accounts"],
        "x-required-scopes": ["accounts:write"],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Platform tujuan, mis. `instagram`, `facebook`, `tiktok`, `youtube`, `linkedin`."
          },
          {
            "name": "redirect",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Redirect URI milik Anda. WAJIB sudah terdaftar di allowlist developer app dan dicocokkan PERSIS (tanpa wildcard, tanpa prefix match). Harus `https` — `http` hanya untuk localhost. Tidak boleh mengandung fragment (`#`) karena akan menelan `code` yang kami tambahkan."
          }
        ],
        "responses": {
          "200": {
            "description": "URL otorisasi platform.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Permintaan tidak valid (validasi schema / input salah).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key tidak ada, tidak valid, atau sudah dicabut.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Plan tidak termasuk api_access (atau api_write untuk endpoint tulis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "API key tidak terikat developer app yang aktif — tidak ada allowlist redirect yang bisa dipercaya.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint tidak ditemukan, atau sumber daya tidak ada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Terlalu banyak permintaan (60/menit per key).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
      "AuthorizeResponse": {
        "type": "object",
        "properties": {
          "authorizeUrl": {
            "type": "string",
            "description": "Arahkan browser pengguna ke URL ini untuk memberi izin."
          }
        },
        "required": ["authorizeUrl"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key format `sk_api_<32 karakter>`. Dibuat di Settings → API. Bisa juga dikirim via header `X-API-Key`."
      }
    }
  }
}